app/requirements.txt
beautifulsoup4==4.12.2
charset-normalizer==2.1.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.21.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
soupsieve==2.5
threadpoolctl==3.1.0
tinydb==4.8.0
Werkzeug==2.2.2
WTForms==2.3.3
flask-wtf
boto3
pydantic